@@ -1,10 +1,10 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.0
 import PackageDescription
 
 let package = Package(
     name: "DreamEchoApp",
     platforms: [
-        .iOS(.v16)
+        .iOS(.v26)
     ],
     products: [
         .library(name: "DreamEchoApp", targets: ["DreamEchoApp"])
